requirements.txt
asgiref==3.3.1
boto3==1.16.51
botocore==1.19.51
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cycler==0.10.0
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jwcrypto==0.8
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.5.1.48
packaging==20.8
pandas==1.1.5
Pillow==8.1.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jwt==3.3.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-learn==0.24.0
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.26.2
